ssrf-flask/app3/requirements.txt
Flask
Werkzeug==2.0.1
Pillow==9.5.0